refactor table into separate file
